functions/trigger-conversion/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-firestore==2.*
cloudevents==1.*
requests==2.*
google-auth==2.*
